Updated Python mapper logic to accommodate required recursion for composed return values (eg. (key,value) tuples) that contain structs
