+++ ksml-runner-axual/ksml-runner.yml
@@ -4,7 +4,7 @@ ksml:
   applicationServerPort: 8080
   workingDirectory: ../examples
   definitions:
-    - 01-demo-inspect.yaml
+#    - 01-demo-inspect.yaml
 #    - 02-demo-copy.yaml
 #    - 03-demo-filter.yaml
 #    - 04-demo-branch.yaml
@@ -16,6 +16,7 @@ ksml:
 #    - 10-demo-queryabletable.yaml
 #    - 11-demo-fieldmodification.yaml
 #    - 12-demo-bytemanipulation.yaml
+    - 13-demo-join.yaml
 
 backend:
   type: axual
